requirements.txt
fastapi
uvicorn
python-dotenv
httpx
stripe
psycopg2-binary
pydantic
